requirements.txt
python-dotenv==0.20.0
loguru==0.6.0

git+https://github.com/boudinfl/pke.git # Keyword Extraction Toolkit

# gRPC deps
grpcio-tools==1.46.3
googleapis-common-protos==1.56.1
grpcio-testing==1.46.3
grpc-interceptor==0.14.1
